refactored the data handlers
moved the parent and subclasses for handlers to handler.py
